requirements.txt
langchain>=0.1.0
faiss-cpu>=1.7.4
arxiv>=2.1.0
numpy>=1.26.0
tqdm>=4.66.0
google-genai>=0.2.0
google-generativeai>=0.3.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
groq>=0.5.0
python-docx>=1.1.0
streamlit>=1.30.0
watchdog>=4.0.0
openai>=1.0.0
